Update index-page.js
Changed Link to a href. 
(Link should only be used for internal links, per the Gatsby documentation.)
Social media links will all be external, by definition, and therefore should use a href.
Urls that use the @ symbol (such as TikTok profiles) will not work with Link, but work fine with a href.